# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tumblr Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/ask/{blog_id}": {
            "post": {
                "tags": [
                    "Ask Blogs"
                ],
                "summary": "ask blog",
                "description": " Primary blog ask another blog",
                "operationId": "askblog",
                "parameters": [
                    {
                        "name": "blog_id",
                        "in": "path",
                        "description": "Blog_id of asking blog ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Question Asked body , flag show if anonymous or not ,true for anonymous ",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_body",
                                    "flag"
                                ],
                                "properties": {
                                    "question_body": {
                                        "type": "string",
                                        "example": "How are you?"
                                    },
                                    "flag": {
                                        "type": "bool",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful  response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "200",
                                                "msg": "ok"
                                            }
                                        },
                                        "response": {
                                            "properties": {
                                                "blog_avatar": {
                                                    "type": "string",
                                                    "example": "/storage/imgname2.extension"
                                                },
                                                "blog_avatar_shape": {
                                                    "type": "string",
                                                    "example": "circle"
                                                },
                                                "blog_username": {
                                                    "type": "string",
                                                    "example": "radwa-ahmed213"
                                                },
                                                "blog_id": {
                                                    "type": "integer",
                                                    "example": 1032
                                                },
                                                "question_body": {
                                                    "type": "string",
                                                    "example": "How are you?"
                                                },
                                                "flag": {
                                                    "type": "bool",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "not found question"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object",
                                            "example": {
                                                "status": "404",
                                                "msg": "Unauthorized"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "security": [
        []
    ]
}